Added firewall interfaces helpers
